requirements.txt
google.generativeai
streamlit
PyPDF2
python-dotenv